ng dev":# Please enter the commit message for your changes. Lines starting

+++ app/src/main/java/com/rapid7/awoods/mission_to_mars/GameObjects/Button.java
@@ -1,8 +1,10 @@
 package com.rapid7.awoods.mission_to_mars.GameObjects;
 
 
+import android.graphics.Canvas;
+
 public abstract class Button extends GameObject {
-    public Button(String image, PositionVector position, String name, double width, double height) {
+    public Button(String image, PositionVector position, String name, float width, float height) {
         super(image, position, name, width, height, true);
     }
