+++ app/src/main/java/com/rapid7/awoods/mission_to_mars/GameObjects/MovingObject.java
@@ -1,13 +1,15 @@
 package com.rapid7.awoods.mission_to_mars.GameObjects;
 
 
+import android.graphics.Canvas;
+
 import java.util.*;
 
 public abstract class MovingObject extends GameObject {
 
-    private double speed;
+    private float speed;
 
-    public MovingObject(String image, PositionVector position, String name, double width, double height, double speed) {
+    public MovingObject(String image, PositionVector position, String name, float width, float height, float speed) {
         super(image, position, name, width, height);
         this.speed = speed;
     }
@@ -15,11 +17,11 @@ public abstract class MovingObject extends GameObject {
     public abstract void update();
 
 
-    public double getSpeed() {
+    public float getSpeed() {
         return speed;
     }
 
-    public void setSpeed(double speed) {
+    public void setSpeed(float speed) {
         this.speed = speed;
     }
 }
