+++ app/src/main/java/com/rapid7/awoods/mission_to_mars/GameObjects/ObjectInstances/Movementbutton.java
@@ -1,5 +1,7 @@
 package com.rapid7.awoods.mission_to_mars.GameObjects.ObjectInstances;
 
+import android.graphics.Canvas;
+
 import com.rapid7.awoods.mission_to_mars.GameObjects.Button;
 import com.rapid7.awoods.mission_to_mars.GameObjects.GameObject;
 import com.rapid7.awoods.mission_to_mars.GameObjects.PositionVector;
@@ -10,7 +12,7 @@ public class Movementbutton extends GameObject{
     Player player;
     boolean moveRight;
 
-    public Movementbutton(String image, PositionVector position, String name, double width, double height,
+    public Movementbutton(String image, PositionVector position, String name, float width, float height,
                           Player player, boolean moveRight) {
         super(image, position, name, width, height);
         this.player = player;
